Relocate TestEnumCompleteness into a dedicated tests/test_enums.py
It lived in test_ipv6_ext.py, a file about IPv6 extension headers,
not enums -- moved verbatim. Also closes the structural gap of
src/netprotocols/_enums.py having no same-named test file of its own.
